+++ src/data_preparation/data_preparation_utils.py
@@ -1,6 +1,7 @@
 """
 The module includes functions for implementing data transformations
 """
+
 # Import Standard Libraries
 import os
 import numpy as np
@@ -12,24 +13,24 @@ from typing import List
 
 # Import Package Modules
 from src.logging_module.logging_module import get_logger
-from src.types import (
+from src.custom_types import (
     EmbeddingsConfig,
     CompressEmbeddingsConfig,
-    EncodingTextConfig
+    EncodingTextConfig,
+    DateExtractionConfig,
 )
 
 # Setup logger
-logger = get_logger(os.path.basename(__file__).split('.')[0],
-                    pathlib.Path(os.getenv('DRUIDIC_GROVE_AI_ROOT_PATH')) /
-                    'src' /
-                    'logging_module' /
-                    'log_configuration.yaml')
+logger = get_logger(
+    os.path.basename(__file__).split(".")[0],
+    pathlib.Path(os.getenv("DRUIDIC_GROVE_AI_ROOT_PATH"))
+    / "src"
+    / "logging_module"
+    / "log_configuration.yaml",
+)
 
 
-def generate_embeddings(
-        texts: List[str],
-        embeddings_config: EmbeddingsConfig
-) -> np.ndarray:
+def generate_embeddings(texts: List[str], embeddings_config: EmbeddingsConfig) -> np.ndarray:
     """
     Generate the embeddings from the input texts through the method
     specified in embeddings_config.method.
@@ -41,33 +42,34 @@ def generate_embeddings(
     Returns:
         sentence_embeddings (numpy.ndarray): Embedded texts (n_samples, embeddings_size)
     """
-    logger.debug('generate_embeddings - Start')
+    logger.debug("generate_embeddings - Start")
 
     # Retrieve embeddings' method
     method = embeddings_config.method
 
     # Switch based on the embeddings' method
     match method:
-        case 'SentenceTransformer':
-            logger.info('generate_embeddings - SentenceTransformer embedding approach')
+        case "SentenceTransformer":
+            logger.info("generate_embeddings - SentenceTransformer embedding approach")
 
             # Instance model
             model = SentenceTransformer(embeddings_config.embedding_model_config.model_name)
 
             # generate embeddings
-            sentence_embeddings = model.encode(texts, convert_to_numpy=embeddings_config.embedding_model_config.numpy_tensor)
+            sentence_embeddings = model.encode(
+                texts, convert_to_numpy=embeddings_config.embedding_model_config.numpy_tensor
+            )
         case _:
-            logger.error('generate_embeddings - Unknown embedding method: %s', method)
-            raise ValueError('Invalid embedding method')
+            logger.error("generate_embeddings - Unknown embedding method: %s", method)
+            raise ValueError("Invalid embedding method")
 
-    logger.debug('generate_embeddings - End')
+    logger.debug("generate_embeddings - End")
 
     return sentence_embeddings
 
 
 def compress_embeddings(
-        input_embeddings: np.ndarray,
-        compress_embeddings_config: CompressEmbeddingsConfig
+    input_embeddings: np.ndarray, compress_embeddings_config: CompressEmbeddingsConfig
 ) -> np.ndarray:
     """
     Compress the input embeddings with the corresponding selected method in
@@ -80,15 +82,15 @@ def compress_embeddings(
     Returns:
         compressed_embeddings (numpy.ndarray): Output embeddings compressed (n_samples, n_components)
     """
-    logger.debug('compress_embeddings - Start')
+    logger.debug("compress_embeddings - Start")
 
     # Retrieve compress method
     method = compress_embeddings_config.method
 
     # Switch based on the compress method
     match method:
-        case 'PCA':
-            logger.info('compress_embeddings - PCA compress approach')
+        case "PCA":
+            logger.info("compress_embeddings - PCA compress approach")
 
             # Instance model
             model = PCA(n_components=compress_embeddings_config.compress_model_config.n_components)
@@ -97,17 +99,17 @@ def compress_embeddings(
             compressed_embeddings = model.fit_transform(input_embeddings)
 
         case _:
-            logger.error('compress_embeddings - Unknown compression method: %s', method)
-            raise ValueError('Invalid compression method')
+            logger.error("compress_embeddings - Unknown compression method: %s", method)
+            raise ValueError("Invalid compression method")
 
-    logger.debug('compress_embeddings - End')
+    logger.debug("compress_embeddings - End")
 
     return compressed_embeddings
 
 
 def encode_text(
-        texts: List[str],
-        config: EncodingTextConfig,
+    texts: List[str],
+    config: EncodingTextConfig,
 ) -> np.ndarray:
     """
     Encode an input text through embeddings and compress their dimensionality.
@@ -119,7 +121,7 @@ def encode_text(
     Returns:
         compressed_embeddings (numpy.ndarray): Output embeddings compressed (n_samples, n_components)
     """
-    logger.debug('encode_text - Start')
+    logger.debug("encode_text - Start")
 
     # Generate embeddings
     embeddings = generate_embeddings(texts, config.embeddings_config)
@@ -127,6 +129,36 @@ def encode_text(
     # Compress embeddings
     compressed_embeddings = compress_embeddings(embeddings, config.compress_embeddings_config)
 
-    logger.debug('encode_text - End')
+    logger.debug("encode_text - End")
 
     return compressed_embeddings
+
+
+def extract_date_information(data: pd.DataFrame, config: DateExtractionConfig) -> pd.DataFrame:
+    """
+    Extract date information from a column included in the ``config.column_name`` like the year, the month, etc.
+
+    Args:
+        data (pd.DataFrame): Input data
+        config (DateExtractionConfig): Configuration including the column_name and date information to extract
+
+    Returns:
+        (pd.DataFrame): Output data with additional columns
+    """
+    logger.debug("extract_date_information - Start")
+
+    # Retrieve column name
+    column_name = config.column_name
+
+    # Convert column to datetime
+    data[column_name] = pd.to_datetime(data[column_name])
+
+    # Extract date information
+    if config.extract_year:
+        data[f"{column_name}_year"] = data[column_name].dt.year
+    if config.extract_month:
+        data[f"{column_name}_month"] = data[column_name].dt.month
+
+    logger.debug("extract_date_information - End")
+
+    return data